move audio store types
